+++ lib/cgo/cipher.crypto.go
@@ -11,47 +11,26 @@ import (
   #include <string.h>
   #include <stdlib.h>
 
-  #include "../../include/skytypes.h"
+  #include "skytypes.h"
 */
 import "C"
 
 //export SKY_cipher_PubKeySlice_Len
-func SKY_cipher_PubKeySlice_Len(_slice *C.cipher__PubKeySlice, _arg0 *int) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	slice := *inplacePubKeySlice(_slice)
-	__arg0 := slice.Len()
-	*_arg0 = __arg0
-	return
+func SKY_cipher_PubKeySlice_Len(_slice *C.cipher__PubKeySlice) int {
+	slice := inplacePubKeySlice(_slice)
+	return slice.Len()
 }
 
 //export SKY_cipher_PubKeySlice_Less
-func SKY_cipher_PubKeySlice_Less(_slice *C.cipher__PubKeySlice, _i, _j int, _arg1 *bool) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	slice := *inplacePubKeySlice(_slice)
-	i := _i
-	j := _j
-	__arg1 := slice.Less(i, j)
-	*_arg1 = __arg1
-	return
+func SKY_cipher_PubKeySlice_Less(_slice *C.cipher__PubKeySlice, _i, _j int) bool {
+	slice := inplacePubKeySlice(_slice)
+	return slice.Less(_i, _j)
 }
 
 //export SKY_cipher_PubKeySlice_Swap
-func SKY_cipher_PubKeySlice_Swap(_slice *C.cipher__PubKeySlice, _i, _j int) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	slice := *inplacePubKeySlice(_slice)
-	i := _i
-	j := _j
-	slice.Swap(i, j)
-	return
+func SKY_cipher_PubKeySlice_Swap(_slice *C.cipher__PubKeySlice, _i, _j int) {
+	slice := inplacePubKeySlice(_slice)
+	slice.Swap(_i, _j)
 }
 
 //export SKY_cipher_RandByte
@@ -67,10 +46,9 @@ func SKY_cipher_RandByte(_n int, _arg1 *C.GoSlice_) (____error_code uint32) {
 }
 
 //export SKY_cipher_NewPubKey
-func SKY_cipher_NewPubKey(_b []byte, _arg1 *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_NewPubKey(_b []byte, _arg1 *C.cipher__PubKey) (errcode uint32) {
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	b := *(*[]byte)(unsafe.Pointer(&_b))
 	__arg1 := cipher.NewPubKey(b)
@@ -79,10 +57,10 @@ func SKY_cipher_NewPubKey(_b []byte, _arg1 *C.cipher__PubKey) (____error_code ui
 }
 
 //export SKY_cipher_MustPubKeyFromHex
-func SKY_cipher_MustPubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_MustPubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (errcode uint32) {
+	errcode = 0
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	s := _s
 	__arg1 := cipher.MustPubKeyFromHex(s)
@@ -91,14 +69,13 @@ func SKY_cipher_MustPubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (____error
 }
 
 //export SKY_cipher_PubKeyFromHex
-func SKY_cipher_PubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_PubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (errcode uint32) {
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	s := _s
 	__arg1, ____return_err := cipher.PubKeyFromHex(s)
-	____error_code = libErrorCode(____return_err)
+	errcode = libErrorCode(____return_err)
 	if ____return_err == nil {
 		copyToBuffer(reflect.ValueOf(__arg1[:]), unsafe.Pointer(_arg1), uint(SizeofPubKey))
 	}
@@ -106,10 +83,9 @@ func SKY_cipher_PubKeyFromHex(_s string, _arg1 *C.cipher__PubKey) (____error_cod
 }
 
 //export SKY_cipher_PubKeyFromSecKey
-func SKY_cipher_PubKeyFromSecKey(_seckey *C.cipher__SecKey, _arg1 *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_PubKeyFromSecKey(_seckey *C.cipher__SecKey, _arg1 *C.cipher__PubKey) (errcode uint32) {
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	seckey := *(*cipher.SecKey)(unsafe.Pointer(_seckey))
 	__arg1 := cipher.PubKeyFromSecKey(seckey)
@@ -118,53 +94,38 @@ func SKY_cipher_PubKeyFromSecKey(_seckey *C.cipher__SecKey, _arg1 *C.cipher__Pub
 }
 
 //export SKY_cipher_PubKeyFromSig
-func SKY_cipher_PubKeyFromSig(_sig *C.cipher__Sig, _hash *C.cipher__SHA256, _arg2 *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	sig := *(*cipher.Sig)(unsafe.Pointer(_sig))
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	__arg2, ____return_err := cipher.PubKeyFromSig(sig, hash)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-		copyToBuffer(reflect.ValueOf(__arg2[:]), unsafe.Pointer(_arg2), uint(SizeofPubKey))
+func SKY_cipher_PubKeyFromSig(_sig *C.cipher__Sig, _hash *C.cipher__SHA256, _arg2 *C.cipher__PubKey) uint32 {
+	sig := (*cipher.Sig)(unsafe.Pointer(_sig))
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+
+	pubkey, err := cipher.PubKeyFromSig(*sig, *hash)
+
+	errcode := libErrorCode(err)
+	if err == nil {
+		copyToBuffer(reflect.ValueOf(pubkey[:]), unsafe.Pointer(_arg2), uint(SizeofPubKey))
+
 	}
-	return
+	return errcode
 }
 
 //export SKY_cipher_PubKey_Verify
-func SKY_cipher_PubKey_Verify(_pk *C.cipher__PubKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	pk := *(*cipher.PubKey)(unsafe.Pointer(_pk))
-	____return_err := pk.Verify()
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_PubKey_Verify(_pk *C.cipher__PubKey) uint32 {
+	pk := (*cipher.PubKey)(unsafe.Pointer(_pk))
+
+	err := pk.Verify()
+	errcode := libErrorCode(err)
+	return errcode
 }
 
 //export SKY_cipher_PubKey_Hex
-func SKY_cipher_PubKey_Hex(_pk *C.cipher__PubKey, _arg0 *C.GoString_) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	pk := *(*cipher.PubKey)(unsafe.Pointer(_pk))
-	__arg0 := pk.Hex()
-	copyString(__arg0, _arg0)
-	return
+func SKY_cipher_PubKey_Hex(_pk *C.cipher__PubKey, _arg1 *C.GoString_) {
+	pk := (*cipher.PubKey)(unsafe.Pointer(_pk))
+	s := pk.Hex()
+	copyString(s, _arg1)
 }
 
 //export SKY_cipher_PubKey_ToAddressHash
-func SKY_cipher_PubKey_ToAddressHash(_pk *C.cipher__PubKey, _arg0 *C.cipher__Ripemd160) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
+func SKY_cipher_PubKey_ToAddressHash(_pk *C.cipher__PubKey, _arg0 *C.cipher__Ripemd160) {
 	pk := (*cipher.PubKey)(unsafe.Pointer(_pk))
 	__arg0 := pk.ToAddressHash()
 	copyToBuffer(reflect.ValueOf(__arg0[:]), unsafe.Pointer(_arg0), uint(SizeofRipemd160))
@@ -172,10 +133,9 @@ func SKY_cipher_PubKey_ToAddressHash(_pk *C.cipher__PubKey, _arg0 *C.cipher__Rip
 }
 
 //export SKY_cipher_NewSecKey
-func SKY_cipher_NewSecKey(_b []byte, _arg1 *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_NewSecKey(_b []byte, _arg1 *C.cipher__SecKey) (errcode uint32) {
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	b := *(*[]byte)(unsafe.Pointer(&_b))
 	__arg1 := cipher.NewSecKey(b)
@@ -196,64 +156,41 @@ func SKY_cipher_MustSecKeyFromHex(_s string, _arg1 *C.cipher__SecKey) (____error
 }
 
 //export SKY_cipher_SecKeyFromHex
-func SKY_cipher_SecKeyFromHex(_s string, _arg1 *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	s := _s
-	__arg1, ____return_err := cipher.SecKeyFromHex(s)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-		copyToBuffer(reflect.ValueOf(__arg1[:]), unsafe.Pointer(_arg1), uint(SizeofSecKey))
+func SKY_cipher_SecKeyFromHex(_s string, _arg1 *C.cipher__SecKey) uint32 {
+	sk, err := cipher.SecKeyFromHex(_s)
+	errcode := libErrorCode(err)
+	if err == nil {
+		copyToBuffer(reflect.ValueOf(sk[:]), unsafe.Pointer(_arg1), uint(SizeofSecKey))
 	}
-	return
+	return errcode
 }
 
 //export SKY_cipher_SecKey_Verify
-func SKY_cipher_SecKey_Verify(_sk *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	sk := *(*cipher.SecKey)(unsafe.Pointer(_sk))
-	____return_err := sk.Verify()
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_SecKey_Verify(_sk *C.cipher__SecKey) uint32 {
+	sk := (*cipher.SecKey)(unsafe.Pointer(_sk))
+	err := sk.Verify()
+	return libErrorCode(err)
 }
 
 //export SKY_cipher_SecKey_Hex
-func SKY_cipher_SecKey_Hex(_sk *C.cipher__SecKey, _arg0 *C.GoString_) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	sk := *(*cipher.SecKey)(unsafe.Pointer(_sk))
-	__arg0 := sk.Hex()
-	copyString(__arg0, _arg0)
-	return
+func SKY_cipher_SecKey_Hex(_sk *C.cipher__SecKey, _arg1 *C.GoString_) {
+	sk := (*cipher.SecKey)(unsafe.Pointer(_sk))
+	s := sk.Hex()
+	copyString(s, _arg1)
 }
 
 //export SKY_cipher_ECDH
-func SKY_cipher_ECDH(_pub *C.cipher__PubKey, _sec *C.cipher__SecKey, _arg2 *C.GoSlice_) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	pub := *(*cipher.PubKey)(unsafe.Pointer(_pub))
-	sec := *(*cipher.SecKey)(unsafe.Pointer(_sec))
-	__arg2 := cipher.ECDH(pub, sec)
-	copyToGoSlice(reflect.ValueOf(__arg2), _arg2)
-	return
+func SKY_cipher_ECDH(_pub *C.cipher__PubKey, _sec *C.cipher__SecKey, _arg2 *C.GoSlice_) {
+	pub := (*cipher.PubKey)(unsafe.Pointer(_pub))
+	sec := (*cipher.SecKey)(unsafe.Pointer(_sec))
+	b := cipher.ECDH(*pub, *sec)
+	copyToGoSlice(reflect.ValueOf(b), _arg2)
 }
 
 //export SKY_cipher_NewSig
-func SKY_cipher_NewSig(_b []byte, _arg1 *C.cipher__Sig) (____error_code uint32) {
-	____error_code = 0
+func SKY_cipher_NewSig(_b []byte, _arg1 *C.cipher__Sig) (errcode uint32) {
 	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
+		errcode = catchApiPanic(errcode, recover())
 	}()
 	b := *(*[]byte)(unsafe.Pointer(&_b))
 	__arg1 := cipher.NewSig(b)
@@ -274,129 +211,79 @@ func SKY_cipher_MustSigFromHex(_s string, _arg1 *C.cipher__Sig) (____error_code 
 }
 
 //export SKY_cipher_SigFromHex
-func SKY_cipher_SigFromHex(_s string, _arg1 *C.cipher__Sig) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	s := _s
-	__arg1, ____return_err := cipher.SigFromHex(s)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-		copyToBuffer(reflect.ValueOf(__arg1[:]), unsafe.Pointer(_arg1), uint(SizeofSig))
+func SKY_cipher_SigFromHex(_s string, _arg1 *C.cipher__Sig) uint32 {
+	s, err := cipher.SigFromHex(_s)
+	errcode := libErrorCode(err)
+	if err == nil {
+		copyToBuffer(reflect.ValueOf(s[:]), unsafe.Pointer(_arg1), uint(SizeofSig))
 	}
-	return
+	return errcode
 }
 
 //export SKY_cipher_Sig_Hex
-func SKY_cipher_Sig_Hex(_s *C.cipher__Sig, _arg0 *C.GoString_) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	s := *(*cipher.Sig)(unsafe.Pointer(_s))
-	__arg0 := s.Hex()
-	copyString(__arg0, _arg0)
-	return
+func SKY_cipher_Sig_Hex(_s *C.cipher__Sig, _arg1 *C.GoString_) {
+	s := (*cipher.Sig)(unsafe.Pointer(_s))
+	copyString(s.Hex(), _arg1)
 }
 
 //export SKY_cipher_SignHash
-func SKY_cipher_SignHash(_hash *C.cipher__SHA256, _sec *C.cipher__SecKey, _arg2 *C.cipher__Sig) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	sec := *(*cipher.SecKey)(unsafe.Pointer(_sec))
-	__arg2 := cipher.SignHash(hash, sec)
-	copyToBuffer(reflect.ValueOf(__arg2[:]), unsafe.Pointer(_arg2), uint(SizeofSig))
-	return
+func SKY_cipher_SignHash(_hash *C.cipher__SHA256, _sec *C.cipher__SecKey, _arg2 *C.cipher__Sig) {
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+	sec := (*cipher.SecKey)(unsafe.Pointer(_sec))
+	s := cipher.SignHash(*hash, *sec)
+	copyToBuffer(reflect.ValueOf(s[:]), unsafe.Pointer(_arg2), uint(SizeofSig))
 }
 
 //export SKY_cipher_ChkSig
-func SKY_cipher_ChkSig(_address *C.cipher__Address, _hash *C.cipher__SHA256, _sig *C.cipher__Sig) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	address := *inplaceAddress(_address)
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	sig := *(*cipher.Sig)(unsafe.Pointer(_sig))
-	____return_err := cipher.ChkSig(address, hash, sig)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_ChkSig(_address *C.cipher__Address, _hash *C.cipher__SHA256, _sig *C.cipher__Sig) uint32 {
+	address := inplaceAddress(_address)
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+	sig := (*cipher.Sig)(unsafe.Pointer(_sig))
+
+	err := cipher.ChkSig(*address, *hash, *sig)
+	return libErrorCode(err)
 }
 
 //export SKY_cipher_VerifySignedHash
-func SKY_cipher_VerifySignedHash(_sig *C.cipher__Sig, _hash *C.cipher__SHA256) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	sig := *(*cipher.Sig)(unsafe.Pointer(_sig))
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	____return_err := cipher.VerifySignedHash(sig, hash)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_VerifySignedHash(_sig *C.cipher__Sig, _hash *C.cipher__SHA256) uint32 {
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+	sig := (*cipher.Sig)(unsafe.Pointer(_sig))
+
+	err := cipher.VerifySignedHash(*sig, *hash)
+	return libErrorCode(err)
 }
 
 //export SKY_cipher_VerifySignature
-func SKY_cipher_VerifySignature(_pubkey *C.cipher__PubKey, _sig *C.cipher__Sig, _hash *C.cipher__SHA256) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	pubkey := *(*cipher.PubKey)(unsafe.Pointer(_pubkey))
-	sig := *(*cipher.Sig)(unsafe.Pointer(_sig))
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	____return_err := cipher.VerifySignature(pubkey, sig, hash)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_VerifySignature(_pubkey *C.cipher__PubKey, _sig *C.cipher__Sig, _hash *C.cipher__SHA256) uint32 {
+	pubkey := (*cipher.PubKey)(unsafe.Pointer(_pubkey))
+	sig := (*cipher.Sig)(unsafe.Pointer(_sig))
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+
+	err := cipher.VerifySignature(*pubkey, *sig, *hash)
+	return libErrorCode(err)
 }
 
 //export SKY_cipher_GenerateKeyPair
-func SKY_cipher_GenerateKeyPair(_arg0 *C.cipher__PubKey, _arg1 *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	__arg0, __arg1 := cipher.GenerateKeyPair()
-	copyToBuffer(reflect.ValueOf(__arg0[:]), unsafe.Pointer(_arg0), uint(SizeofPubKey))
-	copyToBuffer(reflect.ValueOf(__arg1[:]), unsafe.Pointer(_arg1), uint(SizeofSecKey))
-	return
+func SKY_cipher_GenerateKeyPair(_arg0 *C.cipher__PubKey, _arg1 *C.cipher__SecKey) {
+	p, s := cipher.GenerateKeyPair()
+	copyToBuffer(reflect.ValueOf(p[:]), unsafe.Pointer(_arg0), uint(SizeofPubKey))
+	copyToBuffer(reflect.ValueOf(s[:]), unsafe.Pointer(_arg1), uint(SizeofSecKey))
 }
 
 //export SKY_cipher_GenerateDeterministicKeyPair
-func SKY_cipher_GenerateDeterministicKeyPair(_seed []byte, _arg1 *C.cipher__PubKey, _arg2 *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	seed := *(*[]byte)(unsafe.Pointer(&_seed))
-	__arg1, __arg2 := cipher.GenerateDeterministicKeyPair(seed)
-	copyToBuffer(reflect.ValueOf(__arg1[:]), unsafe.Pointer(_arg1), uint(SizeofPubKey))
-	copyToBuffer(reflect.ValueOf(__arg2[:]), unsafe.Pointer(_arg2), uint(SizeofSecKey))
-	return
+func SKY_cipher_GenerateDeterministicKeyPair(_seed []byte, _arg1 *C.cipher__PubKey, _arg2 *C.cipher__SecKey) {
+	p, s := cipher.GenerateDeterministicKeyPair(_seed)
+	copyToBuffer(reflect.ValueOf(p[:]), unsafe.Pointer(_arg1), uint(SizeofPubKey))
+	copyToBuffer(reflect.ValueOf(s[:]), unsafe.Pointer(_arg2), uint(SizeofSecKey))
 }
 
 //export SKY_cipher_DeterministicKeyPairIterator
-func SKY_cipher_DeterministicKeyPairIterator(_seed []byte, _arg1 *C.GoSlice_, _arg2 *C.cipher__PubKey, _arg3 *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	seed := *(*[]byte)(unsafe.Pointer(&_seed))
-	__arg1, __arg2, __arg3 := cipher.DeterministicKeyPairIterator(seed)
-	copyToGoSlice(reflect.ValueOf(__arg1), _arg1)
-	copyToBuffer(reflect.ValueOf(__arg2[:]), unsafe.Pointer(_arg2), uint(SizeofPubKey))
-	copyToBuffer(reflect.ValueOf(__arg3[:]), unsafe.Pointer(_arg3), uint(SizeofSecKey))
-	return
+func SKY_cipher_DeterministicKeyPairIterator(_seed []byte, _arg1 *C.GoSlice_, _arg2 *C.cipher__PubKey, _arg3 *C.cipher__SecKey) {
+	h, p, s := cipher.DeterministicKeyPairIterator(_seed)
+
+	copyToGoSlice(reflect.ValueOf(h), _arg1)
+	copyToBuffer(reflect.ValueOf(p[:]), unsafe.Pointer(_arg2), uint(SizeofPubKey))
+	copyToBuffer(reflect.ValueOf(s[:]), unsafe.Pointer(_arg3), uint(SizeofSecKey))
 }
 
 //export SKY_cipher_GenerateDeterministicKeyPairs
@@ -427,30 +314,18 @@ func SKY_cipher_GenerateDeterministicKeyPairsSeed(_seed []byte, _n int, _arg2 *C
 }
 
 //export SKY_cipher_TestSecKey
-func SKY_cipher_TestSecKey(_seckey *C.cipher__SecKey) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	seckey := *(*cipher.SecKey)(unsafe.Pointer(_seckey))
-	____return_err := cipher.TestSecKey(seckey)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_TestSecKey(_seckey *C.cipher__SecKey) uint32 {
+	seckey := (*cipher.SecKey)(unsafe.Pointer(_seckey))
+
+	err := cipher.TestSecKey(*seckey)
+	return libErrorCode(err)
 }
 
 //export SKY_cipher_TestSecKeyHash
-func SKY_cipher_TestSecKeyHash(_seckey *C.cipher__SecKey, _hash *C.cipher__SHA256) (____error_code uint32) {
-	____error_code = 0
-	defer func() {
-		____error_code = catchApiPanic(____error_code, recover())
-	}()
-	seckey := *(*cipher.SecKey)(unsafe.Pointer(_seckey))
-	hash := *(*cipher.SHA256)(unsafe.Pointer(_hash))
-	____return_err := cipher.TestSecKeyHash(seckey, hash)
-	____error_code = libErrorCode(____return_err)
-	if ____return_err == nil {
-	}
-	return
+func SKY_cipher_TestSecKeyHash(_seckey *C.cipher__SecKey, _hash *C.cipher__SHA256) uint32 {
+	seckey := (*cipher.SecKey)(unsafe.Pointer(_seckey))
+	hash := (*cipher.SHA256)(unsafe.Pointer(_hash))
+
+	err := cipher.TestSecKeyHash(*seckey, *hash)
+	return libErrorCode(err)
 }